update import, remove duplicate libs
