--- api/src/lib/constants.ts
@@ -1,10 +1,0 @@
-// JWT stuff
-export const JWT_SECRET_KEY = process.env.INFINITY_KEYS_SECRET
-export const IK_ID_COOKIE = 'ik-id'
-export const IK_CLAIMS_NAMESPACE = 'https://infinitykeys.io'
-
-// Cloudinary
-export const CLOUDINARY_CLOUD_NAME = 'infinity-keys'
-
-// Sendgrid
-export const SENDGRID_SENDER_ACCOUNT = 'noreply@infinitykeys.io'
